Change type from decimal to float for moneys fields

+++ db/migrate/20161106155803_create_steam_market_orders_details.rb
@@ -2,8 +2,8 @@ class CreateSteamMarketOrdersDetails < ActiveRecord::Migration[5.0]
   def change
     create_table :steam_market_orders_details do |t|
       t.string :item_nameid, null: false, index: true
-      t.decimal :highest_buy_order, default: 0
-      t.decimal :lowest_sell_order, default: 0
+      t.float :highest_buy_order, default: 0
+      t.float :lowest_sell_order, default: 0
       t.integer :buy_order_summary, default: 0
       t.integer :sell_order_summary, default: 0
